@@ -25,5 +25,6 @@ std::string cFlowUniqueID::ToString() const
 cFlowUniqueID cFlowUniqueID::UnverifiedGenerate()
 {
 	std::uniform_int_distribution<uint64_t> dist(1);   // generating in [1,~0u]: cannot include 0, because that is the invalid id.
-	return cFlowUniqueID(dist(std::random_device()));
+	std::random_device rd;
+	return cFlowUniqueID(dist(rd));
 }
